added loading animation to menus, accounts, orders && minor style changes on add page

+++ src/pages/admin/menus.js
@@ -2,7 +2,6 @@ import Link from "next/link";
 import classes from "./menus.module.css";
 import Table from "../../UI/Table/Table";
 import axios from "axios";
-// axios.defaults.baseURL = "http://localhost:3000";
 import {Fragment, useEffect, useMemo, useState} from "react";
 import TableContainer from "@/UI/TableContainer/TableContainer";
 
@@ -10,6 +9,7 @@ import Modal from "../../UI/Modal/Modal";
 import {toast} from "react-toastify";
 import {useRouter} from "next/router";
 import {useSession} from "next-auth/react";
+import Loading from "@/components/Loading/Loading";
 const Menus = (props) => {
   const {data: session, status} = useSession();
 
@@ -17,8 +17,10 @@ const Menus = (props) => {
   const [showModal, setShowModal] = useState(false);
   const [currentItem, setCurrentItem] = useState(null);
   const [menus, setMenus] = useState([]);
+  const [loading, setLoading] = useState(true);
 
   const fetchData = async () => {
+    setLoading(true);
     try {
       let response = await axios.get(`/api/menus/`, {
         withCredentials: true,
@@ -33,6 +35,7 @@ const Menus = (props) => {
         console.log(e);
       }
     }
+    setLoading(false);
   };
 
   useEffect(() => {
@@ -126,6 +129,7 @@ const Menus = (props) => {
     if (status === "loading") {
       return;
     }
+
     if (status === "unauthenticated") {
       router.push("/signin");
       return;
@@ -137,47 +141,48 @@ const Menus = (props) => {
     }
   }, [session]);
 
+  if (loading && session && session.admin) {
+    return <Loading />;
+  }
+
   if (session && session.admin) {
     return (
-      <Fragment>
-        <TableContainer>
-          {showModal && (
-            <Modal showModal={showModal} setShowModal={setShowModal}>
-              <h2 className={classes.test}>
-                Are you sure want to delete <span>{`${currentItem.name}`}</span>{" "}
-                ?
-              </h2>
-              <div style={{whiteSpace: "nowrap"}}>
-                <button
-                  className={classes["sub-btn"]}
-                  onClick={() => toggleModal()}
-                >
-                  Cancel
-                </button>
-                <button
-                  className={classes["delete-btn"]}
-                  onClick={() => deleteItem()}
-                >
-                  Confirm
-                </button>
-              </div>
-            </Modal>
-          )}
-          <div className={classes["table-wrapper"]}>
-            <div className={classes["page-header"]}>
-              <div className={classes["page-title"]}>Menus</div>
-              <Link className={classes["add-btn"]} href="add-menu">
-                Add
-              </Link>
+      <TableContainer>
+        {showModal && (
+          <Modal showModal={showModal} setShowModal={setShowModal}>
+            <h2 className={classes.test}>
+              Are you sure want to delete <span>{`${currentItem.name}`}</span> ?
+            </h2>
+            <div style={{whiteSpace: "nowrap"}}>
+              <button
+                className={classes["sub-btn"]}
+                onClick={() => toggleModal()}
+              >
+                Cancel
+              </button>
+              <button
+                className={classes["delete-btn"]}
+                onClick={() => deleteItem()}
+              >
+                Confirm
+              </button>
             </div>
-            {menus ? (
-              <Table columns={columns} data={[...data]} />
-            ) : (
-              <h1>Loading</h1>
-            )}
+          </Modal>
+        )}
+        <div className={classes["table-wrapper"]}>
+          <div className={classes["page-header"]}>
+            <div className={classes["page-title"]}>Menus</div>
+            <Link className={classes["add-btn"]} href="add-menu">
+              Add
+            </Link>
           </div>
-        </TableContainer>
-      </Fragment>
+          {menus ? (
+            <Table columns={columns} data={[...data]} />
+          ) : (
+            <h1>Loading</h1>
+          )}
+        </div>
+      </TableContainer>
     );
   }
 };
